pi_requirements.txt
python-dotenv
opencv-python-headless
mediapipe
flask
pychromecast
schedule
pygame
scikit-learn
numpy
google-genai
requests
fastapi
uvicorn
protobuf==3.20.3
